Move common things to utils.py
